refactor(debate): use direct ai-client DEFAULT_TEMPERATURE imports (t/2146)
Replace the constants.ts re-export of DEFAULT_TEMPERATURE (added in t/2127)
with direct imports from lib/ai-client/defaults.js in each file that needs
it. Fold the two turnPipeline import lines into one inline-type import to
stay at the 1500-line ESLint ceiling in debateEngine.ts.

Files changed:
- constants.ts: remove DEFAULT_TEMPERATURE re-export
- debateEngine.ts: import from ../ai-client/defaults.js; fold turnPipeline
- debateEngine/phases/crossRespond.ts: import from ../../../ai-client/defaults.js
- explorationSummary.ts: replace local const; import from ../ai-client/defaults.js
- calibrationLogger/extract.ts: import and use DEFAULT_TEMPERATURE (draft_temperature)
